fixed tests

+++ StockSimulator/src/hts/projekt/server/PriceUpdateService.java
@@ -5,13 +5,13 @@ import hts.projekt.shared.Equity;
 public class PriceUpdateService {
 
 	public static void updatePrice(Equity equity) {
-		Double oldPrice = equity.getPrice();
-		Double newPrice = calculateNewPrice(oldPrice);
+		Integer oldPrice = equity.getPrice();
+		Integer newPrice = calculateNewPrice(oldPrice);
 
 		DatabaseConnector.updatePrice(equity.getEquityId(), newPrice);
 	}
 
-	private static Double calculateNewPrice(Double oldPrice) {
-		return Math.random();
+	private static Integer calculateNewPrice(Integer oldPrice) {
+		return (int) Math.random();
 	}
 }
